* 添加 setTimeout, setInterval, clearTimeout, clearInterval * 添加 Promise

+++ utils/os.h
@@ -9,7 +9,7 @@
 
 time_t getTimeInSecond();
 
-uint32_t getTickCount();
+int64_t getTickCount();
 
 void Sleep(uint32_t milliseconds);
